add POST operation, refactor jobs endpoint
Before, the jobs endpoint returned one resource object with an array of
jobs. This worked for basic operations to get jobs, but adding a new job
requires a different JSON strucuture. To conform to JSON API spec, each
job is represented by one resource object which allows to POST a new job
object. PUT and maybe DELETE would be implemented in a future ticket.