scrap getbyid
there is no single ID for a job, it's a combination of posn and suffix. I didn't want to make one up either, so it's going to be null for now. The self link still returns a single job, but it instead uses the posn and suffix instead of a arbitrary jobID